@@ -1,6 +1,8 @@
 {
   "name": "metamodels/attribute_alias",
   "description": "MetaModels attribute for alias",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
   "keywords": [
     "contao",
     "metamodels",
@@ -8,9 +10,6 @@
     "alias",
     "nontranslatable"
   ],
-  "type": "contao-bundle",
-  "homepage": "https://now.metamodel.me/",
-  "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "Christian Schiffler",
@@ -24,6 +23,7 @@
       "role": "Developer"
     }
   ],
+  "homepage": "https://now.metamodel.me/",
   "support": {
     "email": "mail@metamodel.me",
     "issues": "https://github.com/MetaModels/attribute_alias/issues",
@@ -32,20 +32,26 @@
     "source": "https://github.com/MetaModels/attribute_alias"
   },
   "require": {
-    "php": "^7.4 || ^8.0",
-    "contao-community-alliance/dc-general": "^2.3@dev",
-    "contao/core-bundle": "^4.13.0, <5.0",
+    "php": "^8.1",
+    "contao-community-alliance/dc-general": "^2.3",
+    "contao/core-bundle": "^4.13.0 <5.0",
     "doctrine/dbal": "^3.3",
     "menatwork/contao-multicolumnwizard-bundle": "^3.4",
-    "metamodels/core": "^2.3@dev",
+    "metamodels/core": "dev-feature/phpcq-2 as 2.3.0",
     "symfony/dependency-injection": "^5.4",
     "symfony/event-dispatcher": "^5.4",
     "symfony/http-kernel": "^5.4"
   },
   "require-dev": {
     "contao/manager-plugin": "^2.1",
-    "phpcq/all-tasks": "^1.2"
+    "phpcq/runner-bootstrap": "^1.0@dev"
   },
+  "repositories": [
+    {
+      "type": "composer",
+      "url": "https://packages.cyberspectrum.de/r"
+    }
+  ],
   "autoload": {
     "psr-4": {
       "MetaModels\\AttributeAliasBundle\\": "src"
@@ -59,17 +65,18 @@
       "MetaModels\\AttributeAliasBundle\\Test\\": "tests/Test"
     }
   },
-  "extra": {
-    "contao-manager-plugin": "MetaModels\\AttributeAliasBundle\\ContaoManager\\Plugin",
-    "branch-alias": {
-      "dev-feature/2.3.0": "2.3.x-dev"
-    }
-  },
   "config": {
     "allow-plugins": {
       "contao-components/installer": false,
-      "contao/manager-plugin": false
+      "contao/manager-plugin": false,
+      "php-http/discovery": true
     },
     "sort-packages": true
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-feature/2.3.0": "2.3.x-dev"
+    },
+    "contao-manager-plugin": "MetaModels\\AttributeAliasBundle\\ContaoManager\\Plugin"
   }
 }
