@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_alias.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2023 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -16,7 +16,7 @@
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2023 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_alias/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -25,6 +25,8 @@ namespace MetaModels\AttributeAliasBundle\EventListener;
 
 use ContaoCommunityAlliance\DcGeneral\Contao\RequestScopeDeterminator;
 use ContaoCommunityAlliance\DcGeneral\Data\ModelId;
+use ContaoCommunityAlliance\DcGeneral\DataDefinition\ContainerInterface;
+use ContaoCommunityAlliance\DcGeneral\InputProviderInterface;
 use MenAtWork\MultiColumnWizardBundle\Event\GetOptionsEvent;
 use MetaModels\Attribute\IInternal;
 use MetaModels\IFactory;
@@ -39,14 +41,14 @@ class GetOptionsListener
      *
      * @var RequestScopeDeterminator
      */
-    private $scopeDeterminator;
+    private RequestScopeDeterminator $scopeDeterminator;
 
     /**
      * Metamodels factory.
      *
      * @var IFactory
      */
-    private $factory;
+    private IFactory $factory;
 
     /**
      * GetOptionsListener constructor.
@@ -73,8 +75,10 @@ class GetOptionsListener
             return false;
         }
 
+        $dataDefinition = $event->getEnvironment()->getDataDefinition();
+        assert($dataDefinition instanceof ContainerInterface);
         return
-            ($event->getEnvironment()->getDataDefinition()->getName() === 'tl_metamodel_attribute')
+            ($dataDefinition->getName() === 'tl_metamodel_attribute')
             && ($event->getPropertyName() === 'alias_fields')
             && ($event->getSubPropertyName() === 'field_attribute');
     }
@@ -88,16 +92,16 @@ class GetOptionsListener
      */
     public function getOptions(GetOptionsEvent $event)
     {
-        if (null !== $event->getOptions() || !$this->isEventForMe($event)) {
+        if (!$this->isEventForMe($event)) {
             return;
         }
 
         $model       = $event->getModel();
         $metaModelId = $model->getProperty('pid');
         if (!$metaModelId) {
-            $metaModelId = ModelId::fromSerialized(
-                $event->getEnvironment()->getInputProvider()->getParameter('pid')
-            )->getId();
+            $inputProvider = $event->getEnvironment()->getInputProvider();
+            assert($inputProvider instanceof InputProviderInterface);
+            $metaModelId = ModelId::fromSerialized($inputProvider->getParameter('pid'))->getId();
         }
 
         $metaModelName = $this->factory->translateIdToMetaModelName($metaModelId);
