@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_alias.
  *
- * (c) 2012-2021 The MetaModels team.
+ * (c) 2012-2023 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,21 +13,23 @@
  * @package    MetaModels/attribute_alias
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2021 The MetaModels team.
+ * @copyright  2012-2023 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_alias/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\AttributeAliasBundle\Migration;
 
 use Contao\CoreBundle\Migration\AbstractMigration;
 use Contao\CoreBundle\Migration\MigrationResult;
 use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\FetchMode;
 use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\ColumnDiff;
+use Doctrine\DBAL\Schema\SchemaException;
 use Doctrine\DBAL\Schema\TableDiff;
 
 /**
@@ -42,7 +44,7 @@ class AllowNullMigration extends AbstractMigration
      *
      * @var Connection
      */
-    private $connection;
+    private Connection $connection;
 
     /**
      * Create a new instance.
@@ -74,7 +76,7 @@ class AllowNullMigration extends AbstractMigration
      */
     public function shouldRun(): bool
     {
-        $schemaManager = $this->connection->getSchemaManager();
+        $schemaManager = $this->connection->createSchemaManager();
 
         if (!$schemaManager->tablesExist(['tl_metamodel', 'tl_metamodel_attribute'])) {
             return false;
@@ -104,7 +106,7 @@ class AllowNullMigration extends AbstractMigration
             }
         }
 
-        return new MigrationResult(true, 'Adjusted column(s): ' . implode(', ', $message));
+        return new MigrationResult(true, 'Adjusted column(s): ' . \implode(', ', $message));
     }
 
     /**
@@ -118,19 +120,17 @@ class AllowNullMigration extends AbstractMigration
         if (empty($langColumns)) {
             return [];
         }
-        $schemaManager = $this->connection->getSchemaManager();
+        $schemaManager = $this->connection->createSchemaManager();
 
         $result = [];
         foreach ($langColumns as $tableName => $tableColumnNames) {
             /** @var Column[] $columns */
             $columns = [];
             // The schema manager return the column list with lowercase keys, wo got to use the real names.
-            \array_map(
-                function (Column $column) use (&$columns) {
-                    $columns[$column->getName()] = $column;
-                },
-                $schemaManager->listTableColumns($tableName)
-            );
+            $table = $schemaManager->introspectTable($tableName);
+            foreach ($table->getColumns() as $column) {
+                $columns[$column->getName()] = $column;
+            }
             foreach ($tableColumnNames as $tableColumnName) {
                 $column = ($columns[$tableColumnName] ?? null);
                 if (null === $column) {
@@ -163,7 +163,7 @@ class AllowNullMigration extends AbstractMigration
             ->leftJoin('attribute', 'tl_metamodel', 'metamodel', 'attribute.pid = metamodel.id')
             ->where('attribute.type=:type')
             ->setParameter('type', 'alias')
-            ->execute()
+            ->executeQuery()
             ->fetchAllAssociative();
 
         $result = [];
@@ -184,28 +184,28 @@ class AllowNullMigration extends AbstractMigration
      * @param Column $column    The column.
      *
      * @return void
+     *
+     * @throws Exception
+     * @throws SchemaException
      */
     private function fixColumn(string $tableName, Column $column): void
     {
-        $manager = $this->connection->getSchemaManager();
-        $table   = $manager->listTableDetails($tableName);
+        $manager = $this->connection->createSchemaManager();
+        $table   = $manager->introspectTable($tableName);
+        $updated = $manager->introspectTable($tableName);
 
-        $changeColumn = new Column($column->getName(), $column->getType());
-        $changeColumn
-            ->setLength($column->getLength())
+        $updated->getColumn($column->getName())
             ->setNotnull(false)
             ->setDefault(null);
-        $columnDiff = new ColumnDiff($column->getName(), $changeColumn);
 
-        $tableDiff                   = new TableDiff($tableName);
-        $tableDiff->fromTable        = $table;
-        $tableDiff->changedColumns[] = $columnDiff;
+        $tableDiff = $manager->createComparator()->compareTables($table, $updated);
+
         $manager->alterTable($tableDiff);
 
         $this->connection->createQueryBuilder()
             ->update($tableName, 't')
             ->set('t.' . $column->getName(), 'null')
             ->where('t.' . $column->getName() . ' = ""')
-            ->execute();
+            ->executeQuery();
     }
 }
