@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_alias.
  *
- * (c) 2012-2018 The MetaModels team.
+ * (c) 2012-2023 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,15 +12,18 @@
  *
  * @package    MetaModels/attribute_alias
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2012-2018 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2023 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_alias/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\AttributeAliasBundle\Test\Schema;
 
+use Doctrine\DBAL\Exception;
+use Doctrine\DBAL\Schema\SchemaException;
 use Doctrine\DBAL\Schema\Table;
 use Doctrine\DBAL\Types\Type;
 use Doctrine\DBAL\Types\Types;
@@ -30,6 +33,8 @@ use PHPUnit\Framework\TestCase;
 
 /**
  * This tests the schema generator.
+ *
+ * @covers \MetaModels\AttributeAliasBundle\Schema\DoctrineSchemaGenerator
  */
 class DoctrineSchemaGeneratorTest extends TestCase
 {
@@ -37,6 +42,10 @@ class DoctrineSchemaGeneratorTest extends TestCase
      * Test the generate method.
      *
      * @return void
+     *
+     * @throws Exception
+     * @throws SchemaException
+     * @throws \ReflectionException
      */
     public function testGenerate(): void
     {
